Finished Refactoring all Bukkit tasks with TimeArbiter.java or SwordScheduler.java methods (with 2 specific exceptions) to reduce bloat and allow for a new, dynamic system that reduces verbosity.

+++ src/main/java/btm/sword/utility/SwordTimeUnit.java
@@ -2,14 +2,14 @@ package btm.sword.utility;
 
 public final class SwordTimeUnit {
     public static final int TICKS_PER_SECOND = 20;
-    public static final long MILLISECONDS_PER_TICK = 50L;
+    public static final int MILLISECONDS_PER_TICK = 50;
 
     // Primary conversions - milliseconds as base unit
     public static int millisToTicks(long millis) {
         return (int) Math.round(millis / (double) MILLISECONDS_PER_TICK);
     }
 
-    public static long ticksToMillis(int ticks) {
+    public static int ticksToMillis(int ticks) {
         return ticks * MILLISECONDS_PER_TICK;
     }
